create global components; fix creating a server
