+++ assets/common/vue/folder_structure_end_state.vue
@@ -1,0 +1,7 @@
+<template lang="pug">
+div
+  h2-hash Folder Structure End State
+  p Upon completing the Steps section below, your project's folder structure should look similar to the following:
+  code-block(:header="false")
+    slot
+</template>
