+++ assets/common/vue/global_components.js
@@ -1,0 +1,16 @@
+import Vue from "vue";
+
+import CodeBlock from "/common/vue/code_block.vue";
+Vue.component("code-block", CodeBlock);
+
+import FolderStructureEndState from "/common/vue/folder_structure_end_state.vue";
+Vue.component("folder-structure-end-state", FolderStructureEndState);
+
+import H2Hash from "/common/vue/h2_hash.vue";
+Vue.component("h2-hash", H2Hash);
+
+import Page from "/common/vue/page.vue";
+Vue.component("page", Page);
+
+import Request from "/common/vue/request.vue";
+Vue.component("request", Request);
